requirements.txt
# Core
numpy==2.2.2
opencv-python==4.10.0.84

# Emotion
fer==22.5.1

# Video handling
moviepy==1.0.3
imageio==2.34.1
decorator==5.1.1

# System
psutil==5.9.8
pywin32==311; platform_system == "Windows"

# Notifications
plyer==2.1.0

# Config
PyYAML==6.0.2

# Optional analytics
pandas==2.2.3
